abstract things

+++ GIT/OOP/lab5/lab5/Rectangle.cs
@@ -2,15 +2,15 @@
 {
     public class Rectangle : Figure
     {
-        private float area;
-        private float perimeter;
-        private float width;
-        private float height;
-        public float Perimeter()
+        private double area;
+        private double perimeter;
+        private double width;
+        private double height;
+        public double Perimeter()
         {
             return 0;
         }
-        public float Area()
+        public double Area()
         {
             return 0;
         }
